chore: move copy to different folder
