+++ src/client/bot-firewall/botFirewallCopy.ts
